requirements.txt
django==3.2
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
requests==2.27.1
geopy==2.2.0
dj-database-url==1.0.0
environs==9.5.0
